Move loadtest CLI handling out of main
